requirements.txt
streamlit==1.32.0
google-cloud-storage==2.16.0
google-generativeai>=0.8.3 
Pillow==10.2.0
python-dotenv==1.0.1